+++ demo/typescript/index.js
@@ -7,7 +7,8 @@
 // demo/typescript/index.ts
 import Human from "../../dist/human.esm.js";
 var humanConfig = {
-  modelBasePath: "../../models"
+  modelBasePath: "../../models",
+  filter: { equalization: false }
 };
 var human = new Human(humanConfig);
 human.env["perfadd"] = false;
@@ -79,8 +80,8 @@ async function drawLoop() {
   setTimeout(drawLoop, 30);
 }
 async function main() {
-  log("human version:", human.version, "tfjs version:", human.tf.version_core);
-  log("platform:", human.env.platform, "agent:", human.env.agent);
+  log("human version:", human.version, "| tfjs version:", human.tf.version_core);
+  log("platform:", human.env.platform, "| agent:", human.env.agent);
   status("loading...");
   await human.load();
   log("backend:", human.tf.getBackend(), "| available:", human.env.backends);