+++ types/src/config.d.ts
@@ -110,7 +110,9 @@ export interface SegmentationConfig extends GenericConfig {
 export interface FilterConfig {
     /** @property are image filters enabled? */
     enabled: boolean;
-    /** @property perform image histogram equalization */
+    /** @property perform image histogram equalization
+     * - equalization is performed on input as a whole and detected face before its passed for further analysis
+    */
     equalization: boolean;
     /** resize input width
     * - if both width and height are set to 0, there is no resizing
@@ -207,6 +209,8 @@ export interface Config {
      * default: 0.7
     */
     cacheSensitivity: number;
+    /** Perform immediate garbage collection on deallocated tensors instead of caching them */
+    deallocate: boolean;
     /** Internal Variable */
     skipAllowed: boolean;
     /** {@link FilterConfig} */