+++ types/src/image/enhance.d.ts
@@ -2,4 +2,4 @@
  * Image enhancements
  */
 import type { Tensor } from '../exports';
-export declare function histogramEqualization(input: Tensor): Tensor;
+export declare function histogramEqualization(inputImage: Tensor): Promise<Tensor>;