+++ types/src/image/image.d.ts
@@ -4,8 +4,8 @@
 import type { Input, AnyCanvas, Tensor, Config } from '../exports';
 export declare function canvas(width: any, height: any): AnyCanvas;
 export declare function copy(input: AnyCanvas, output?: AnyCanvas): AnyCanvas;
-export declare function process(input: Input, config: Config, getTensor?: boolean): {
+export declare function process(input: Input, config: Config, getTensor?: boolean): Promise<{
     tensor: Tensor | null;
     canvas: AnyCanvas | null;
-};
+}>;
 export declare function skip(config: any, input: Tensor): Promise<boolean>;